clinic-admin-backend/requirements.txt
# Core Python Version: 3.13.3

# Backend API - FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# Database - MongoDB Driver
motor==3.3.2
pymongo==4.6.0

# Data Validation and Settings
pydantic==2.11.7
pydantic-settings==2.1.0
pydantic-core==2.33.2

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
cryptography==42.0.0

# HTTP and File Handling
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6

# Configuration and Environment
python-dotenv==1.0.0

# Email Validation
email-validator==2.2.0

# Admin UI (Optional - for admin interface)
streamlit==1.28.1
streamlit-authenticator==0.2.3
plotly==5.17.0
pandas==2.1.4

# Development and Code Quality
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Additional Dependencies (Auto-detected from current environment)
anyio==3.7.1
certifi==2025.7.14
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==3.0.2
packaging==25.0
sniffio==1.3.1
typing-extensions==4.14.1
watchfiles==1.1.0
websockets==15.0.1